requirements.txt
playwright
gspread
oauth2client
python-dotenv
requests
